kkjj-30 <refactor> 구매 가격 자료형 변경

+++ entity/AirPurchase.java
@@ -20,7 +20,7 @@ public class AirPurchase extends BasePurchase {
     @OneToMany(mappedBy = "airPurchase", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<AirPassenger> airPassengers = new ArrayList<>();
 
-    public AirPurchase(SeatClass seatClass, Integer price, Long member_id, String name, String number, String email, LocalDateTime purchaseDate, LocalDateTime paymentDueDate) {
+    public AirPurchase(SeatClass seatClass, Long price, Long member_id, String name, String number, String email, LocalDateTime purchaseDate, LocalDateTime paymentDueDate) {
         super(price, member_id, name, number, email, purchaseDate, paymentDueDate);
         this.seatClass = seatClass;
